+++ yt/frontends/sph/fields.py
@@ -25,10 +25,13 @@ from yt.fields.field_info_container import \
 
 from .definitions import \
     gadget_ptypes, \
-    ghdf5_ptypes
+    ghdf5_ptypes,\
+    eaglenetwork_ion_lookup
 
+from yt.units.yt_array import YTQuantity
 from yt.config import ytcfg
 from yt.utilities.physical_constants import mh
+from yt.utilities.periodic_table import periodic_table
 from yt.fields.species_fields import \
     add_species_field_by_fraction, \
     add_species_field_by_density, \
@@ -91,31 +94,31 @@ class SPHFieldInfo(FieldInfoContainer):
         setup_species_fields(self, ptype)
 
 class TipsyFieldInfo(SPHFieldInfo):
-
+    aux_particle_fields = {
+        'uDotFB':("uDotFB", ("code_mass * code_velocity**2", ["uDotFB"], None)),
+        'uDotAV':("uDotAV", ("code_mass * code_velocity**2", ["uDotAV"], None)),
+        'uDotPdV':("uDotPdV", ("code_mass * code_velocity**2", ["uDotPdV"], None)),
+        'uDotHydro':("uDotHydro", ("code_mass * code_velocity**2", ["uDotHydro"], None)),
+        'uDotDiff':("uDotDiff", ("code_mass * code_velocity**2", ["uDotDiff"], None)),
+        'uDot':("uDot", ("code_mass * code_velocity**2", ["uDot"], None)),
+        'coolontime':("coolontime", ("code_time", ["coolontime"], None)),
+        'timeform':("timeform", ("code_time", ["timeform"], None)),
+        'massform':("massform", ("code_mass", ["massform"], None)),
+        'HI':("HI", ("dimensionless", ["HI"], None)),
+        'HII':("HII", ("dimensionless", ["HII"], None)),
+        'HeI':("HeI", ("dimensionless", ["HeI"], None)),
+        'HeII':("HeII", ("dimensionless", ["HeII"], None)),
+        'OxMassFrac':("OxMassFrac", ("dimensionless", ["OxMassFrac"], None)),
+        'FeMassFrac':("FeMassFrac", ("dimensionless", ["FeMassFrac"], None)),
+        'c':("c", ("code_velocity", ["c"], None)),
+        'acc':("acc", ("code_velocity / code_time", ["acc"], None)),
+        'accg':("accg", ("code_velocity / code_time", ["accg"], None))}
+    
     def __init__(self, ds, field_list, slice_info = None):
-        aux_particle_fields = {
-                'uDotFB':("uDotFB", ("code_mass * code_velocity**2", ["uDotFB"], None)),
-                'uDotAV':("uDotAV", ("code_mass * code_velocity**2", ["uDotAV"], None)),
-                'uDotPdV':("uDotPdV", ("code_mass * code_velocity**2", ["uDotPdV"], None)),
-                'uDotHydro':("uDotHydro", ("code_mass * code_velocity**2", ["uDotHydro"], None)),
-                'uDotDiff':("uDotDiff", ("code_mass * code_velocity**2", ["uDotDiff"], None)),
-                'uDot':("uDot", ("code_mass * code_velocity**2", ["uDot"], None)),
-                'coolontime':("coolontime", ("code_time", ["coolontime"], None)),
-                'timeform':("timeform", ("code_time", ["timeform"], None)),
-                'massform':("massform", ("code_mass", ["massform"], None)),
-                'HI':("HI", ("dimensionless", ["HI"], None)),
-                'HII':("HII", ("dimensionless", ["HII"], None)),
-                'HeI':("HeI", ("dimensionless", ["HeI"], None)),
-                'HeII':("HeII", ("dimensionless", ["HeII"], None)),
-                'OxMassFrac':("OxMassFrac", ("dimensionless", ["OxMassFrac"], None)),
-                'FeMassFrac':("FeMassFrac", ("dimensionless", ["FeMassFrac"], None)),
-                'c':("c", ("code_velocity", ["c"], None)),
-                'acc':("acc", ("code_velocity / code_time", ["acc"], None)),
-                'accg':("accg", ("code_velocity / code_time", ["accg"], None))}
         for field in field_list:
-            if field[1] in aux_particle_fields.keys() and \
-                aux_particle_fields[field[1]] not in self.known_particle_fields:
-                self.known_particle_fields += (aux_particle_fields[field[1]],)
+            if field[1] in self.aux_particle_fields.keys() and \
+                self.aux_particle_fields[field[1]] not in self.known_particle_fields:
+                self.known_particle_fields += (self.aux_particle_fields[field[1]],)
         super(TipsyFieldInfo,self).__init__(ds, field_list, slice_info)
 
 
@@ -188,8 +191,14 @@ class OWLSFieldInfo(SPHFieldInfo):
             ftype='gas'
         elif ptype == 'PartType1':
             ftype='dm'
+        elif ptype == 'PartType2':
+            ftype='PartType2'
+        elif ptype == 'PartType3':
+            ftype='PartType3'
         elif ptype == 'PartType4':
             ftype='star'
+        elif ptype == 'PartType5':
+            ftype='BH'
         elif ptype == 'all':
             ftype='all'
         
@@ -408,3 +417,54 @@ class OWLSFieldInfo(SPHFieldInfo):
             raise RuntimeError, "Failed to download owls ion data."
 
         return owls_ion_path
+
+
+class EagleNetworkFieldInfo(OWLSFieldInfo):
+
+    _ions = \
+        ('H1', 'H2', 'He1', 'He2','He3', 'C1',\
+         'C2', 'C3', 'C4', 'C5', 'C6', 'C7', 'N1', 'N2', \
+         'N3', 'N4', 'N5', 'N6', 'N7', 'N8', 'O1', 'O2', 'O3', \
+         'O4', 'O5', 'O6', 'O7', 'O8', 'O9', 'Ne1', 'Ne2',\
+         'Ne3', 'Ne4', 'Ne5', 'Ne6', 'Ne7', 'Ne8', 'Ne9', 'Ne10',\
+         'Ne11', 'Mg1', 'Mg2', 'Mg3', 'Mg4', 'Mg5', 'Mg6', 'Mg7',\
+         'Mg8', 'Mg9', 'Mg10', 'Mg11', 'Mg12', 'Mg13', 'Si1', 'Si2',\
+         'Si3', 'Si4', 'Si5', 'Si6', 'Si7', 'Si8', 'Si9', 'Si10',\
+         'Si11', 'Si12', 'Si13', 'Si14', 'Si15', 'Si16', 'Si17',\
+         'Ca1', 'Ca2', 'Ca3', 'Ca4', 'Ca5', 'Ca6', 'Ca7', 'Ca8',\
+         'Ca9', 'Ca10', 'Ca11', 'Ca12', 'Ca13', 'Ca14', 'Ca15',\
+         'Ca16', 'Ca17', 'Ca18', 'Ca19', 'Ca20', 'Ca21', 'Fe1',\
+         'Fe2', 'Fe3', 'Fe4', 'Fe5', 'Fe6', 'Fe7', 'Fe8', 'Fe9',\
+         'Fe10', 'Fe11', 'Fe12', 'Fe13', 'Fe14', 'Fe15', 'Fe16',\
+         'Fe17', 'Fe18', 'Fe19', 'Fe20', 'Fe21', 'Fe22', 'Fe23',\
+         'Fe24', 'Fe25', 'Fe25', 'Fe27',)
+
+    def __init__(self, *args, **kwargs):
+        
+        super(EagleNetworkFieldInfo,self).__init__( *args, **kwargs )
+        
+    def _create_ion_density_func( self, ftype, ion ):
+        """ returns a function that calculates the ion density of a particle. 
+        """ 
+
+        def _ion_density(field, data):
+
+            # Lookup the index of the ion 
+            index = eaglenetwork_ion_lookup[ion] 
+
+            # Ion to hydrogen number density ratio
+            ion_chem = data[ftype, "Chemistry_%03i"%index]
+
+            # Mass of a single ion
+            if ion[0:2].isalpha():
+                symbol = ion[0:2].capitalize()
+            else:
+                symbol = ion[0:1].capitalize()
+            m_ion = YTQuantity(periodic_table.elements_by_symbol[symbol].weight, 'amu')
+
+            # hydrogen number density 
+            n_H = data["PartType0", "H_number_density"] 
+
+            return m_ion*ion_chem*n_H 
+        
+        return _ion_density